Reuse shared YARN rotary helpers and consolidate CP slice helpers
Drop DeepSeek-V4 CSA's local _yarn_find_correction_dim/_yarn_linear_ramp_mask
copies in favor of the shared rotary utilities; the YARN correction range is
identical on this path (beta_fast != beta_slow), so forward output is unchanged.

Move the generic contiguous CP-slice validators (local_position_ids_for_cp,
local_sequence_tensor_for_cp) out of the attention module into the canonical
parallel/cp primitive, leaving only attention-specific compressed-CP helpers in
the attention layer; redirect the protocol to import them from parallel/cp.
